+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### master
 
+* bug fixes
+  * remove fonts from broccoli pipeline
+  
 ### 1.0.2 - 2017-04-14
 
 * bug fixes